requirements.txt
Flask
flask-socketio
opencv-python
numpy
Pillow